Finnaly UnityBuild for xrGame

+++ src/xrGame/WeaponMagazined.cpp
@@ -1247,7 +1247,7 @@ void	CWeaponMagazined::OnNextFireMode		()
 {
 	if (!m_bHasDifferentFireModes) return;
 	if (GetState() != eIdle) return;
-	m_iCurFireMode = (m_iCurFireMode+1+m_aFireModes.size()) % m_aFireModes.size();
+	m_iCurFireMode = (m_iCurFireMode+1+m_aFireModes.size()) % (int)m_aFireModes.size();
 	SetQueueSize(GetCurrentFireMode());
 };
 
@@ -1255,7 +1255,7 @@ void	CWeaponMagazined::OnPrevFireMode		()
 {
 	if (!m_bHasDifferentFireModes) return;
 	if (GetState() != eIdle) return;
-	m_iCurFireMode = (m_iCurFireMode-1+m_aFireModes.size()) % m_aFireModes.size();
+	m_iCurFireMode = (m_iCurFireMode-1+m_aFireModes.size()) % (int)m_aFireModes.size();
 	SetQueueSize(GetCurrentFireMode());	
 };
 
@@ -1345,7 +1345,7 @@ bool CWeaponMagazined::GetBriefInfo( II_BriefInfo& info )
 	GetSuitableAmmoTotal();//update m_BriefInfo_CalcFrame
 	info.grenade				= "";
 
-	u32 at_size = m_ammoTypes.size();
+	u32 at_size = (u32)m_ammoTypes.size();
 	if ( unlimited_ammo() || at_size == 0 )
 	{
 		info.fmj_ammo._set( "--" );
