+++ src/xrGame/sound_player.cpp
@@ -83,7 +83,7 @@ u32 CSoundPlayer::add				(LPCSTR prefix, u32 max_count, ESoundTypes type, u32 pr
 	VERIFY								(pair.first == (CSoundCollectionParams&)sound_params);
 	VERIFY								(pair.second);
 	m_sounds.insert						(std::make_pair(internal_type,std::make_pair(sound_params,pair.second)));
-	return								(pair.second->m_sounds.size());
+	return								(u32)(pair.second->m_sounds.size());
 }
 
 void CSoundPlayer::remove			(u32 internal_type)
@@ -295,13 +295,13 @@ const ref_sound &CSoundPlayer::CSoundCollection::random	(const u32 &id)
 	}
 
 	if (m_sounds.size() <= 2) {
-		m_last_sound_id		= CRandom32::random(m_sounds.size());
+		m_last_sound_id		= CRandom32::random((u32)m_sounds.size());
 		return				(*m_sounds[m_last_sound_id]);
 	}
 	
 	u32						result;
 	do {
-		result				= CRandom32::random(m_sounds.size());
+		result				= CRandom32::random((u32)m_sounds.size());
 	}
 	while (result == m_last_sound_id);
